requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-talisman==1.1.0
requests==2.31.0
pycryptodome==3.20.0
gunicorn==21.2.0
